Now project use discrete representation of brightness (0...255) instead float value (0.0...1.0).

+++ image-triangulation/TExtension.cs
@@ -55,7 +55,7 @@ namespace image_triangulation
             {
                 while (tmpFileReader.BaseStream.Position != tmpFileReader.BaseStream.Length)
                 {
-                    Pixel pixel = new Pixel(tmpFileReader.ReadInt16(), tmpFileReader.ReadInt16(), tmpFileReader.ReadSingle());
+                    Pixel pixel = new Pixel(tmpFileReader.ReadInt16(), tmpFileReader.ReadInt16(), tmpFileReader.ReadByte());
                     pivotPoints.Add(pixel);
                 }
             }
